Changes to improve GPS/sensor polling, linear regression math, and glide ratio calculation
* GPS updates now use single update for better control over update
frequency (requires API 9)
* Fixed several cases where update MS is treated as second values
* Wiped out LocationUtils distance calculation in favor of Android OS
one (old one seemed incorrect, and produced huge numbers for glide
ratio)
* Upped bariometric sensor polling frequency (gives linearregression
more samples)
* Test GPS writer now attempts to read the test file in correct timing
* Updated linear regression math to reduce number magnitude mismatch and
multiplication of raw timestamps

+++ src/com/geeksville/location/AndroidBarometerClient.java
@@ -147,7 +147,7 @@ public class AndroidBarometerClient extends SensorClient implements
 	 */
 	@Override
 	public float getVerticalSpeed() {
-		return regression.getSlope() * 1000;
+		return (float)(regression.getSlope() * 1000);
 	}
 
 	@Override
