+++ src/com/geeksville/location/FlynetBarometerClient.java
@@ -143,7 +143,7 @@ public class FlynetBarometerClient extends Observable implements
   @Override
   public float getVerticalSpeed() {
     try {
-      return regression.getSlope() * 1000;
+      return (float)(regression.getSlope() * 1000);
     } catch (ArithmeticException divByZero) {
       return Float.NaN;
     }
